requirements.txt
flask==3.0.2
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
email-validator==2.2.0
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
python-dotenv==1.0.1
gunicorn==21.2.0
